converter/pyproject.toml
[project]
name = "image-converter"
version = "0.1.0"
description = "Converter between custom binary image format and PNG"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.scripts]
image-converter = "converter.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/converter"]
